+++ src/main/java/com/commercehub/gradle/plugin/avro/DefaultAvroExtension.java
@@ -25,14 +25,18 @@ public class DefaultAvroExtension implements AvroExtension {
     }
 
     @Override
-    public String getFieldVisibility() { return fieldVisibility; }
+    public String getFieldVisibility() {
+        return fieldVisibility;
+    }
     
     public void setFieldVisibility(String fieldVisibility) {
         this.fieldVisibility = fieldVisibility;
     }
 
     @Override
-    public String getTemplateDirectory() { return templateDirectory; }
+    public String getTemplateDirectory() {
+        return templateDirectory;
+    }
 
     public void setTemplateDirectory(String templateDirectory) {
         this.templateDirectory = templateDirectory;