+++ src/test/groovy/com/commercehub/gradle/plugin/avro/AvroPluginFunctionalSpec.groovy
@@ -1,7 +1,5 @@
 package com.commercehub.gradle.plugin.avro
 
-import org.gradle.testkit.runner.GradleRunner
-
 import java.nio.file.Files
 
 import static org.gradle.testkit.runner.TaskOutcome.FAILED
@@ -114,6 +112,7 @@ class AvroPluginFunctionalSpec extends FunctionalSpec {
         then:
         result.task(":generateAvroJava").outcome == FAILED
         result.standardError.contains("> Could not compile schema definition files:")
-        result.standardError.contains("* src/main/avro/enumMalformed.avsc: \"enum\" is not a defined name. The type of the \"gender\" field must be a defined name or a {\"type\": ...} expression.")
+        result.standardError.contains("* src/main/avro/enumMalformed.avsc: \"enum\" is not a defined name. The type of the \"gender\" " +
+                "field must be a defined name or a {\"type\": ...} expression.")
     }
 }